[FIX] rename major file
