+++ functions/api/routes/major/majorListUniversityGET.js
